app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

langchain
langchain-core
langchain-community

langchain-groq

sentence-transformers
langchain-huggingface

redis[search]==4.6.0

PyGithub

gunicorn